Reorganize tests into their own files
