+++ test/asyncPosIntegratedControllerTests.cpp
@@ -1,0 +1,44 @@
+/**
+ * @author Ryan Benasutti, WPI
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+#include "okapi/api/control/async/asyncPosIntegratedController.hpp"
+#include "test/tests/api/implMocks.hpp"
+#include <gtest/gtest.h>
+
+using namespace okapi;
+
+class AsyncPosIntegratedControllerTest : public ::testing::Test {
+  protected:
+  void SetUp() override {
+    motor = new MockMotor();
+    controller =
+      new AsyncPosIntegratedController(std::shared_ptr<MockMotor>(motor), createTimeUtil());
+  }
+
+  void TearDown() override {
+    delete controller;
+  }
+
+  MockMotor *motor;
+  AsyncPosIntegratedController *controller;
+};
+
+TEST_F(AsyncPosIntegratedControllerTest, SettledWhenDisabled) {
+  assertControllerIsSettledWhenDisabled(*controller, 100.0);
+}
+
+TEST_F(AsyncPosIntegratedControllerTest, WaitUntilSettledWorksWhenDisabled) {
+  assertWaitUntilSettledWorksWhenDisabled(*controller);
+}
+
+TEST_F(AsyncPosIntegratedControllerTest, FollowsDisableLifecycle) {
+  assertControllerFollowsDisableLifecycle(*controller, motor->lastPosition, motor->lastVoltage);
+}
+
+TEST_F(AsyncPosIntegratedControllerTest, FollowsTargetLifecycle) {
+  assertControllerFollowsTargetLifecycle(*controller);
+}
