+++ test/controllerTests.cpp
@@ -26,61 +26,6 @@
 
 using namespace okapi;
 
-template <typename I, typename O>
-void assertControllerIsSettledWhenDisabled(ClosedLoopController<I, O> &controller, I target) {
-  controller.flipDisable(false);
-  controller.setTarget(target);
-  EXPECT_FALSE(controller.isSettled());
-
-  controller.flipDisable(true);
-  EXPECT_TRUE(controller.isSettled());
-}
-
-template <typename I, typename O>
-void assertWaitUntilSettledWorksWhenDisabled(AsyncController<I, O> &controller) {
-  controller.flipDisable(true);
-  controller.waitUntilSettled();
-}
-
-void assertControllerFollowsDisableLifecycle(AsyncController<double, double> &controller,
-                                             std::int16_t &domainValue,
-                                             std::int16_t &voltageValue) {
-  EXPECT_FALSE(controller.isDisabled()) << "Should not be disabled at the start.";
-
-  controller.setTarget(100);
-  EXPECT_EQ(domainValue, 100) << "Should be on by default.";
-
-  controller.flipDisable();
-  EXPECT_TRUE(controller.isDisabled()) << "Should be disabled after flipDisable";
-  EXPECT_EQ(voltageValue, 0) << "Disabling the controller should turn the motor off";
-
-  controller.flipDisable();
-  EXPECT_FALSE(controller.isDisabled()) << "Should not be disabled after flipDisable";
-  EXPECT_EQ(domainValue, 100)
-    << "Re-enabling the controller should move the motor to the previous target";
-
-  controller.flipDisable();
-  EXPECT_TRUE(controller.isDisabled()) << "Should be disabled after flipDisable";
-  controller.reset();
-  EXPECT_TRUE(controller.isDisabled()) << "Should be disabled after reset";
-  EXPECT_EQ(voltageValue, 0) << "Resetting the controller should not change the current target";
-
-  controller.flipDisable();
-  EXPECT_FALSE(controller.isDisabled()) << "Should not be disabled after flipDisable";
-  domainValue = 1337;            // Sample value to check it doesn't change
-  MockRate().delayUntil(100_ms); // Wait for it to possibly change
-  EXPECT_EQ(domainValue, 1337)
-    << "Re-enabling the controller after a reset should not move the motor";
-}
-
-void assertControllerFollowsTargetLifecycle(AsyncController<double, double> &controller) {
-  EXPECT_DOUBLE_EQ(0, controller.getError()) << "Should start with 0 error";
-  controller.setTarget(100);
-  EXPECT_DOUBLE_EQ(100, controller.getError());
-  controller.setTarget(0);
-  EXPECT_DOUBLE_EQ(0, controller.getError());
-}
-
 class IterativeControllerWithSimulatorTest : public ::testing::Test {
   protected:
   virtual void SetUp() {
@@ -145,40 +90,6 @@ TEST_F(IterativeControllerWithSimulatorTest, IterativeVelPIDControllerFeedForwar
   }
 }
 
-class IterativePosPIDControllerTest : public ::testing::Test {
-  protected:
-  virtual void SetUp() {
-    controller = new IterativePosPIDController(0.1, 0, 0, 0, createConstantTimeUtil(10_ms));
-  }
-
-  virtual void TearDown() {
-    delete controller;
-  }
-
-  IterativePosPIDController *controller;
-};
-
-TEST_F(IterativePosPIDControllerTest, BasicKpTest) {
-  EXPECT_DOUBLE_EQ(controller->step(1), 0.1 * -1);
-}
-
-TEST_F(IterativePosPIDControllerTest, DefaultErrorBounds_ErrorOfZero) {
-  EXPECT_DOUBLE_EQ(controller->step(0), 0);
-}
-
-TEST_F(IterativePosPIDControllerTest, DefaultErrorBounds_ErrorOfOne) {
-  EXPECT_DOUBLE_EQ(controller->step(1), 0.1 * -1);
-}
-
-TEST_F(IterativePosPIDControllerTest, DefaultErrorBounds_LargeError) {
-  const double largeError = 10000000000000;
-  EXPECT_DOUBLE_EQ(controller->step(largeError), -1);
-}
-
-TEST_F(IterativePosPIDControllerTest, SettledWhenDisabled) {
-  assertControllerIsSettledWhenDisabled(*controller, 100.0);
-}
-
 TEST(IterativeMotorVelocityControllerTest, IterativeMotorVelocityController) {
   class MockIterativeVelPIDController : public IterativeVelPIDController {
     public:
@@ -216,114 +127,6 @@ TEST(IterativeMotorVelocityControllerTest, IterativeMotorVelocityController) {
 
   controller.step(-0.5);
   EXPECT_NEAR(motor->lastVelocity, -63, 0.01);
-}
-
-class AsyncPosIntegratedControllerTest : public ::testing::Test {
-  protected:
-  void SetUp() override {
-    motor = new MockMotor();
-    controller =
-      new AsyncPosIntegratedController(std::shared_ptr<MockMotor>(motor), createTimeUtil());
-  }
-
-  void TearDown() override {
-    delete controller;
-  }
-
-  MockMotor *motor;
-  AsyncPosIntegratedController *controller;
-};
-
-TEST_F(AsyncPosIntegratedControllerTest, SettledWhenDisabled) {
-  assertControllerIsSettledWhenDisabled(*controller, 100.0);
-}
-
-TEST_F(AsyncPosIntegratedControllerTest, WaitUntilSettledWorksWhenDisabled) {
-  assertWaitUntilSettledWorksWhenDisabled(*controller);
-}
-
-TEST_F(AsyncPosIntegratedControllerTest, FollowsDisableLifecycle) {
-  assertControllerFollowsDisableLifecycle(*controller, motor->lastPosition, motor->lastVoltage);
-}
-
-TEST_F(AsyncPosIntegratedControllerTest, FollowsTargetLifecycle) {
-  assertControllerFollowsTargetLifecycle(*controller);
-}
-
-class AsyncVelIntegratedControllerTest : public ::testing::Test {
-  protected:
-  void SetUp() override {
-    motor = new MockMotor();
-    controller =
-      new AsyncVelIntegratedController(std::shared_ptr<MockMotor>(motor), createTimeUtil());
-  }
-
-  void TearDown() override {
-    delete controller;
-  }
-
-  MockMotor *motor;
-  AsyncVelIntegratedController *controller;
-};
-
-TEST_F(AsyncVelIntegratedControllerTest, SettledWhenDisabled) {
-  assertControllerIsSettledWhenDisabled(*controller, 100.0);
-}
-
-TEST_F(AsyncVelIntegratedControllerTest, WaitUntilSettledWorksWhenDisabled) {
-  assertWaitUntilSettledWorksWhenDisabled(*controller);
-}
-
-TEST_F(AsyncVelIntegratedControllerTest, FollowsDisableLifecycle) {
-  assertControllerFollowsDisableLifecycle(*controller, motor->lastVelocity, motor->lastVoltage);
-}
-
-TEST_F(AsyncVelIntegratedControllerTest, FollowsTargetLifecycle) {
-  assertControllerFollowsTargetLifecycle(*controller);
-}
-
-class IterativeVelPIDControllerTest : public ::testing::Test {
-  protected:
-  void SetUp() override {
-    controller = new IterativeVelPIDController(
-      0,
-      0,
-      0,
-      0,
-      std::make_unique<VelMath>(
-        1800, std::make_shared<PassthroughFilter>(), std::make_unique<ConstantMockTimer>(10_ms)),
-      createTimeUtil(Supplier<std::unique_ptr<AbstractTimer>>(
-        []() { return std::make_unique<ConstantMockTimer>(10_ms); })));
-  }
-
-  void TearDown() override {
-    delete controller;
-  }
-
-  IterativeVelPIDController *controller;
-};
-
-TEST_F(IterativeVelPIDControllerTest, SettledWhenDisabled) {
-  controller->setGains(0.1, 0.1, 0.1, 0.1);
-  assertControllerIsSettledWhenDisabled(*controller, 100.0);
-}
-
-TEST_F(IterativeVelPIDControllerTest, StaticFrictionGainUsesTargetSign) {
-  controller->setGains(0, 0, 0, 0.1);
-
-  controller->setTarget(1);
-  EXPECT_DOUBLE_EQ(controller->step(0), 1 * 0.1);
-
-  // Use the same target but send the error to 0 to make sure the gain is applied to the target and
-  // not the error
-  EXPECT_DOUBLE_EQ(controller->step(1), 1 * 0.1);
-
-  controller->setTarget(-1);
-  EXPECT_DOUBLE_EQ(controller->step(0), -1 * 0.1);
-
-  // Use the same target but send the error to 0 to make sure the gain is applied to the target and
-  // not the error
-  EXPECT_DOUBLE_EQ(controller->step(-1), -1 * 0.1);
 }
 
 TEST(FilteredControllerInputTest, InputShouldBePassedThrough) {
@@ -386,76 +189,4 @@ TEST(SettledUtilTest, ZeroTime) {
   EXPECT_FALSE(settledUtil.isSettled(60));
   EXPECT_FALSE(settledUtil.isSettled(55));
   EXPECT_TRUE(settledUtil.isSettled(50));
-}
-
-class AsyncMotionProfileControllerTest : public ::testing::Test {
-  protected:
-  void SetUp() override {
-    leftMotor = new MockMotor();
-    rightMotor = new MockMotor();
-
-    model = new SkidSteerModel(std::unique_ptr<AbstractMotor>(leftMotor),
-                               std::unique_ptr<AbstractMotor>(rightMotor));
-
-    controller = new AsyncMotionProfileController(
-      createTimeUtil(), 1.0, 2.0, 10.0, std::shared_ptr<SkidSteerModel>(model), 10.5_in);
-    controller->startThread();
-  }
-
-  void TearDown() override {
-    delete controller;
-  }
-
-  MockMotor *leftMotor;
-  MockMotor *rightMotor;
-  SkidSteerModel *model;
-  AsyncMotionProfileController *controller;
-};
-
-TEST_F(AsyncMotionProfileControllerTest, SettledWhenDisabled) {
-  assertControllerIsSettledWhenDisabled(*controller, std::string("A"));
-}
-
-TEST_F(AsyncMotionProfileControllerTest, WaitUntilSettledWorksWhenDisabled) {
-  assertWaitUntilSettledWorksWhenDisabled(*controller);
-}
-
-TEST_F(AsyncMotionProfileControllerTest, MotorsAreStoppedAfterSettling) {
-  controller->generatePath({Point{0_m, 0_m, 0_deg}, Point{3_ft, 0_m, 45_deg}}, "A");
-
-  controller->setTarget("A");
-  controller->waitUntilSettled();
-
-  assertMotorsHaveBeenStopped(leftMotor, rightMotor);
-  EXPECT_GT(leftMotor->maxVelocity, 0);
-  EXPECT_GT(rightMotor->maxVelocity, 0);
-}
-TEST_F(AsyncMotionProfileControllerTest, WrongPathNameDoesNotMoveAnything) {
-  controller->setTarget("A");
-  controller->waitUntilSettled();
-
-  EXPECT_EQ(leftMotor->maxVelocity, 0);
-  EXPECT_EQ(rightMotor->maxVelocity, 0);
-}
-
-TEST_F(AsyncMotionProfileControllerTest, TwoPathsOverwriteEachOther) {
-  controller->generatePath({Point{0_m, 0_m, 0_deg}, Point{3_ft, 0_m, 45_deg}}, "A");
-  controller->generatePath({Point{0_m, 0_m, 0_deg}, Point{3_ft, 2_ft, 45_deg}}, "A");
-
-  controller->setTarget("A");
-  controller->waitUntilSettled();
-  assertMotorsHaveBeenStopped(leftMotor, rightMotor);
-  EXPECT_GT(leftMotor->maxVelocity, 0);
-  EXPECT_GT(rightMotor->maxVelocity, 0);
-}
-
-TEST_F(AsyncMotionProfileControllerTest, ImpossiblePathThrowsException) {
-  EXPECT_THROW(controller->generatePath({Point{0_m, 0_m, 0_deg},
-                                         Point{3_ft, 0_m, 0_deg},
-                                         Point{3_ft, 1_ft, 0_deg},
-                                         Point{2_ft, 1_ft, 0_deg},
-                                         Point{1_ft, 1_m, 0_deg},
-                                         Point{1_ft, 0_m, 0_deg}},
-                                        "A"),
-               std::runtime_error);
 }
