+++ test/xDriveModelTests.cpp
@@ -5,35 +5,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include "okapi/api/chassis/model/skidSteerModel.hpp"
-#include "okapi/api/chassis/model/threeEncoderSkidSteerModel.hpp"
 #include "okapi/api/chassis/model/xDriveModel.hpp"
-#include "okapi/api/device/motor/abstractMotor.hpp"
 #include "test/tests/api/implMocks.hpp"
 #include <gtest/gtest.h>
 
 using namespace okapi;
-
-void assertMotorsGearsetEquals(const AbstractMotor::gearset expected,
-                               const std::initializer_list<MockMotor> &motors) {
-  for (auto &motor : motors) {
-    EXPECT_EQ(expected, motor.gearset);
-  }
-}
-
-void assertMotorsBrakeModeEquals(const AbstractMotor::brakeMode expected,
-                                 const std::initializer_list<MockMotor> &motors) {
-  for (auto &motor : motors) {
-    EXPECT_EQ(expected, motor.brakeMode);
-  }
-}
-
-void assertMotorsEncoderUnitsEquals(const AbstractMotor::encoderUnits expected,
-                                    const std::initializer_list<MockMotor> &motors) {
-  for (auto &motor : motors) {
-    EXPECT_EQ(expected, motor.encoderUnits);
-  }
-}
 
 class XDriveModelTest : public ::testing::Test {
   public:
@@ -293,201 +269,4 @@ TEST_F(XDriveModelTest, SetEncoderUnitsTest) {
   assertMotorsEncoderUnitsEquals(
     AbstractMotor::encoderUnits::counts,
     {*topLeftMotor, *topRightMotor, *bottomRightMotor, *bottomLeftMotor});
-}
-
-class SkidSteerModelTest : public ::testing::Test {
-  public:
-  SkidSteerModelTest() : model(leftMotor, rightMotor, 127) {
-  }
-
-  void assertAllMotorsLastVelocity(const std::int16_t expectedLastVelocity) const {
-    EXPECT_EQ(leftMotor->lastVelocity, expectedLastVelocity);
-    EXPECT_EQ(rightMotor->lastVelocity, expectedLastVelocity);
-  }
-
-  void assertAllMotorsLastVoltage(const std::int16_t expectedLastVoltage) const {
-    EXPECT_EQ(leftMotor->lastVoltage, expectedLastVoltage);
-    EXPECT_EQ(rightMotor->lastVoltage, expectedLastVoltage);
-  }
-
-  void assertLeftAndRightMotorsLastVelocity(const std::int16_t expectedLeftLastVelocity,
-                                            const std::int16_t expectedRightLastVelocity) const {
-    EXPECT_EQ(leftMotor->lastVelocity, expectedLeftLastVelocity);
-    EXPECT_EQ(rightMotor->lastVelocity, expectedRightLastVelocity);
-  }
-
-  void assertLeftAndRightMotorsLastVoltage(const std::int16_t expectedLeftLastVoltage,
-                                           const std::int16_t expectedRightLastVoltage) const {
-    EXPECT_EQ(leftMotor->lastVoltage, expectedLeftLastVoltage);
-    EXPECT_EQ(rightMotor->lastVoltage, expectedRightLastVoltage);
-  }
-
-  std::shared_ptr<MockMotor> leftMotor = std::make_shared<MockMotor>();
-  std::shared_ptr<MockMotor> rightMotor = std::make_shared<MockMotor>();
-  SkidSteerModel model;
-};
-
-TEST_F(SkidSteerModelTest, ForwardHalfPower) {
-  model.forward(0.5);
-  assertAllMotorsLastVelocity(63);
-}
-
-TEST_F(SkidSteerModelTest, ForwardBoundsInput) {
-  model.forward(10);
-  assertAllMotorsLastVelocity(127);
-}
-
-TEST_F(SkidSteerModelTest, RotateHalfPower) {
-  model.rotate(0.5);
-  assertLeftAndRightMotorsLastVelocity(63, -63);
-}
-
-TEST_F(SkidSteerModelTest, RotateBoundsInput) {
-  model.rotate(10);
-  assertLeftAndRightMotorsLastVelocity(127, -127);
-}
-
-TEST_F(SkidSteerModelTest, DriveVectorHalfPower) {
-  model.driveVector(0.25, 0.25);
-  assertLeftAndRightMotorsLastVelocity(63, 0);
-}
-
-TEST_F(SkidSteerModelTest, DriveVectorBoundsInput) {
-  model.driveVector(0.9, 0.25);
-  assertLeftAndRightMotorsLastVelocity(127, 71);
-}
-
-TEST_F(SkidSteerModelTest, StopTest) {
-  leftMotor->lastVelocity = 100;
-  rightMotor->lastVelocity = 100;
-
-  model.stop();
-
-  assertAllMotorsLastVelocity(0);
-}
-
-TEST_F(SkidSteerModelTest, LeftHalfPower) {
-  model.left(0.5);
-  assertLeftAndRightMotorsLastVelocity(63, 0);
-}
-
-TEST_F(SkidSteerModelTest, RightHalfPower) {
-  model.right(0.5);
-  assertLeftAndRightMotorsLastVelocity(0, 63);
-}
-
-TEST_F(SkidSteerModelTest, TankHalfPower) {
-  model.tank(0.5, 0.5);
-
-  assertAllMotorsLastVelocity(0);
-  assertAllMotorsLastVoltage(63);
-}
-
-TEST_F(SkidSteerModelTest, TankBoundsInput) {
-  model.tank(10, 10);
-
-  assertAllMotorsLastVelocity(0);
-  assertAllMotorsLastVoltage(127);
-}
-
-TEST_F(SkidSteerModelTest, TankThresholds) {
-  model.tank(0.1, 0.1, 0.3);
-
-  assertAllMotorsLastVelocity(0);
-  assertAllMotorsLastVoltage(0);
-}
-
-TEST_F(SkidSteerModelTest, ArcadeHalfPower) {
-  model.arcade(0.5, 0);
-
-  assertAllMotorsLastVelocity(0);
-  assertAllMotorsLastVoltage(63);
-}
-
-TEST_F(SkidSteerModelTest, ArcadeHalfPowerTurn) {
-  model.arcade(0, 0.5);
-
-  assertAllMotorsLastVelocity(0);
-  assertLeftAndRightMotorsLastVoltage(63, -63);
-}
-
-TEST_F(SkidSteerModelTest, ArcadeBoundsInput) {
-  model.arcade(10, 0);
-
-  assertAllMotorsLastVelocity(0);
-  assertAllMotorsLastVoltage(127);
-}
-
-TEST_F(SkidSteerModelTest, ArcadeThresholds) {
-  model.arcade(0.2, 0, 0.5);
-
-  assertAllMotorsLastVelocity(0);
-  assertAllMotorsLastVoltage(0);
-}
-
-TEST_F(SkidSteerModelTest, SetGearsetTest) {
-  model.setGearing(AbstractMotor::gearset::green);
-  assertMotorsGearsetEquals(AbstractMotor::gearset::green, {*leftMotor, *rightMotor});
-}
-
-TEST_F(SkidSteerModelTest, SetBrakeModeTest) {
-  model.setBrakeMode(AbstractMotor::brakeMode::hold);
-  assertMotorsBrakeModeEquals(AbstractMotor::brakeMode::hold, {*leftMotor, *rightMotor});
-}
-
-TEST_F(SkidSteerModelTest, SetEncoderUnitsTest) {
-  model.setEncoderUnits(AbstractMotor::encoderUnits::counts);
-  assertMotorsEncoderUnitsEquals(AbstractMotor::encoderUnits::counts, {*leftMotor, *rightMotor});
-}
-
-class ThreeEncoderSkidSteerModelTest : public ::testing::Test {
-  protected:
-  void SetUp() override {
-    leftMotor = std::make_shared<MockMotor>();
-    rightMotor = std::make_shared<MockMotor>();
-    leftSensor = std::make_shared<MockContinuousRotarySensor>();
-    rightSensor = std::make_shared<MockContinuousRotarySensor>();
-    middleSensor = std::make_shared<MockContinuousRotarySensor>();
-    model =
-      new ThreeEncoderSkidSteerModel(leftMotor, rightMotor, leftSensor, middleSensor, rightSensor);
-  }
-
-  void TearDown() override {
-    delete model;
-  }
-
-  std::shared_ptr<MockMotor> leftMotor;
-  std::shared_ptr<MockMotor> rightMotor;
-  std::shared_ptr<MockContinuousRotarySensor> leftSensor;
-  std::shared_ptr<MockContinuousRotarySensor> rightSensor;
-  std::shared_ptr<MockContinuousRotarySensor> middleSensor;
-  ThreeEncoderSkidSteerModel *model;
-};
-
-TEST_F(ThreeEncoderSkidSteerModelTest, GetSensorValsIsCompatibleWithSkidSteerModel) {
-  leftSensor->value = 1;
-  rightSensor->value = 2;
-  middleSensor->value = 3;
-
-  auto skidSteerModelVals =
-    SkidSteerModel(leftMotor, rightMotor, leftSensor, rightSensor).getSensorVals();
-  auto threeEncoderSkidSteerModelVals = model->getSensorVals();
-
-  EXPECT_EQ(skidSteerModelVals[0], threeEncoderSkidSteerModelVals[0]);
-  EXPECT_EQ(skidSteerModelVals[1], threeEncoderSkidSteerModelVals[1]);
-}
-
-TEST_F(ThreeEncoderSkidSteerModelTest, GetSensorValsIsCompatibleWithXDriveModel) {
-  leftSensor->value = 1;
-  rightSensor->value = 2;
-  middleSensor->value = 3;
-
-  // nonsense motors because it doesn't matter
-  auto xDriveModelVals =
-    XDriveModel(leftMotor, rightMotor, leftMotor, rightMotor, leftSensor, rightSensor)
-      .getSensorVals();
-  auto threeEncoderSkidSteerModelVals = model->getSensorVals();
-
-  EXPECT_EQ(xDriveModelVals[0], threeEncoderSkidSteerModelVals[0]);
-  EXPECT_EQ(xDriveModelVals[1], threeEncoderSkidSteerModelVals[1]);
 }
